refactor(appContext): refactor app context
